fix: mindmap issue (#7335)
### Change
- Mindmap style and issue
- Shape text rendering
- Add extra `padding` and `shadow` yfield

### Detail
#### Shape text rendering
The shape text always shifts when entering the editing mode. The reason is that the text is rendering on `ideographic` baseline not `alphabetic` which is used in CSS rendering. So, to align with the DOM rendering, the baseline has changed to `alphabetic` and the start vertical position is the ascent and line gap of the font.