fix(blocks): ime compatibility of slash menu (#7452)
Closes:  #7451

## What changes
Use the `beforeInput` event instead of the `keyDown` event, as the latter cannot capture the actual input character when IME are active. For example, when pressing <kbd>/</kbd> with Sogou IME, the actual character input is `、`, but at this time `event.key` is equal to `Process`.

After using the `beforeInput` event, there is no need to check whether various modifier keys are pressed, such as in non-English keyboard layouts, like the French keyboard (`Shift + Slash = /`) and the German keyboard (`Shift + 7 = /`).

However, in the new implementation with Chinese IME actived, pressing<kbd> \ </kbd> key also tirgger slash menu becasue of `、` inputed.

## Test cases
- [x] Update a playwright test cases for '、'
- [x] Test on `fr` and `de` keyboard layout manually
- [x] Test on Microsoft Pinyin and Sougou Pinyin manually
- [x] Test mobile browser manually
  - [x] Andorid Chrome (known issue: can not close slash menu by pressing backspace)
  - [x] IOS Chrome & Safari