fix(blocks): preserve block ids when pasting (#7386)
This closes #6366, closes BS-139.

This PR merges part of snapshot to the `fromPointState.model`. The first snapshot is deleted after being merged to the `fromPointState.model`, so the snapshot could probably be an empty array.
We delete snapshot from snapshots instead of block from blocks mainly because that if blocks are deleted immediately after they are created will cause errors as operations in `connectcallback` does not have any corresponding blocks attached.