fix(blocks): switch to `file-type` v19.1.0 (#7523)
This pull request is for sunsetting the `@sgtpooki/file-type` and using the official file-type v19.1.0.

The `file-type` package has replaced the usage of `node:buffer` with `Uint8Array`. This change allows it to run safely in the browser now.

The file-type package is used in the attachment-block and image-block modules to determine the file type of uploaded files.

Related to https://github.com/toeverything/blocksuite/pull/6803 https://github.com/sindresorhus/file-type/issues/578#event-13411340875