fix: no need to query portal when convert edgeless block to canvas (#8040)
To fix: [PD-1632](https://linear.app/affine-design/issue/PD-1632/edgeless-make-it-real-报错)

The to canvas function is still relied on in some whiteboard-related AI functions like make it real and needs to be fixed before switching to the new to png method.

In this pr, remove edgeless portal query and block query function.