feat: add file type detection when exporting (#8425)
This pull request enhances the `AssetsManager` to support file type detection.

Currently, images are named as xxxxx.blob when exporting markdown, which is not user-friendly. This PR improves the export experience by introducing `file-type` to guess the file type.

Before

<img width="329" alt="Screenshot 2024-09-21 at 21 15 18" src="https://github.com/user-attachments/assets/987eb720-7abe-42ed-b1e2-709e81aa003f">

After

<img width="632" alt="Screenshot 2024-09-23 at 18 49 02" src="https://github.com/user-attachments/assets/bf8c55ab-762a-4344-aec8-f9588608dfaf">

<img width="359" alt="Screenshot 2024-09-23 at 18 48 48" src="https://github.com/user-attachments/assets/6681670f-23ed-4d53-b25b-58de57b08498">

---

Note that the `MemoryBlobSource` used by the playground will preserve the MIME type of the blob, which is different from the BlobSource of AFFiNE. Tweak the `MemoryBlobSource` to test these changes.

```diff
// packages/framework/sync/src/blob/impl/memory.ts

  set(key: string, value: Blob) {
+    // this.map.set(key, value);
+    this.map.set(key, new Blob([value]));
    return Promise.resolve(key);
  }
```